move entity dependency logic store
